+++ internal/domain/feed.go
@@ -1,9 +1,5 @@
 package domain
 
-import (
-	"sync"
-)
-
 type MarketPrice struct {
 	Market Market
 	Price  Price
@@ -12,7 +8,7 @@ type MarketPrice struct {
 // Feed represents a source of MarketPrice data
 type Feed interface {
 	AddMarketPrice(marketPrice MarketPrice)
-	getMarketPriceChan() <-chan MarketPrice
+	GetMarketPriceChan() <-chan MarketPrice
 }
 
 type feed struct {
@@ -31,30 +27,6 @@ func (f feed) AddMarketPrice(marketPrice MarketPrice) {
 	f.marketPriceChan <- marketPrice
 }
 
-// merge gathers several feeds into a unique channel
-func merge(feeds ...Feed) <-chan MarketPrice {
-	mergedChan := make(chan MarketPrice)
-	var wg sync.WaitGroup
-
-	wg.Add(len(feeds))
-	for _, feed := range feeds {
-		c := feed.getMarketPriceChan()
-		go func(c <-chan MarketPrice) {
-			for marketPrice := range c {
-				mergedChan <- marketPrice
-			}
-			wg.Done()
-		}(c)
-	}
-
-	go func() {
-		wg.Wait()
-		close(mergedChan)
-	}()
-
-	return mergedChan
-}
-
-func (f feed) getMarketPriceChan() <-chan MarketPrice {
+func (f feed) GetMarketPriceChan() <-chan MarketPrice {
 	return f.marketPriceChan
 }
